Removed subtracting burnedSeigniorageAmount from total supply
Seigniorage is burnt, but WTON still exist in Deposit manager. Therefore it is locked as opposed to being burnt.

main.js file contained logic that subtracts burned seigniorage amount, this part has been removed.

Total supply for price dashboard should be equal to the google sheet now as we do not account for burned seignorage